move newsletter workflows to new lib
Co-authored by: David Blatcher <dv.bltchr@gmail.com>
Co-authored by: Idris Elliott <idrise@davidjelliott.com>
