+++ libs/newsletter-workflows/README.md
@@ -1,0 +1,11 @@
+# newsletter-workflows
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build newsletter-workflows` to build the library.
+
+## Running unit tests
+
+Run `nx test newsletter-workflows` to execute the unit tests via [Jest](https://jestjs.io).
