+++ libs/newsletter-workflows/jest.config.ts
@@ -1,0 +1,16 @@
+/* eslint-disable */
+export default {
+	displayName: 'newsletter-workflows',
+	preset: '../../jest.preset.js',
+	globals: {
+		'ts-jest': {
+			tsconfig: '<rootDir>/tsconfig.spec.json',
+		},
+	},
+	testEnvironment: 'node',
+	transform: {
+		'^.+\\.[tj]s$': 'ts-jest',
+	},
+	moduleFileExtensions: ['ts', 'js', 'html'],
+	coverageDirectory: '../../coverage/libs/newsletter-workflows',
+};
