+++ libs/newsletter-workflows/package.json
@@ -1,0 +1,5 @@
+{
+	"name": "@newsletters-nx/newsletter-workflows",
+	"version": "0.0.1",
+	"type": "commonjs"
+}
