+++ libs/newsletter-workflows/project.json
@@ -1,0 +1,34 @@
+{
+	"name": "newsletter-workflows",
+	"$schema": "../../node_modules/nx/schemas/project-schema.json",
+	"sourceRoot": "libs/newsletter-workflows/src",
+	"projectType": "library",
+	"targets": {
+		"build": {
+			"executor": "@nrwl/js:tsc",
+			"outputs": ["{options.outputPath}"],
+			"options": {
+				"outputPath": "dist/libs/newsletter-workflows",
+				"main": "libs/newsletter-workflows/src/index.ts",
+				"tsConfig": "libs/newsletter-workflows/tsconfig.lib.json",
+				"assets": ["libs/newsletter-workflows/*.md"]
+			}
+		},
+		"lint": {
+			"executor": "@nrwl/linter:eslint",
+			"outputs": ["{options.outputFile}"],
+			"options": {
+				"lintFilePatterns": ["libs/newsletter-workflows/**/*.ts"]
+			}
+		},
+		"test": {
+			"executor": "@nrwl/jest:jest",
+			"outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+			"options": {
+				"jestConfig": "libs/newsletter-workflows/jest.config.ts",
+				"passWithNoTests": true
+			}
+		}
+	},
+	"tags": []
+}
