+++ libs/newsletter-workflows/src/lib/newsletter-workflows.spec.ts
@@ -1,0 +1,7 @@
+import { newsletterWorkflows } from './newsletter-workflows';
+
+describe('newsletterWorkflows', () => {
+	// it('should work', () => {
+	// 	expect(newsletterWorkflows()).toEqual('newsletter-workflows');
+	// });
+});
